+++ src/Services/NutritionService/GymApp.NutritionService.Data/Entities/Food.cs
@@ -6,10 +6,10 @@ public class Food
 {
     public int Id { get; set; }
     public string? Name { get; set; }
-    public int Calories { get; set; }
-    public int Protein { get; set; }
-    public int Carbohydrates { get; set; }
-    public int Fats { get; set; }
+    public double Calories { get; set; }
+    public double Protein { get; set; }
+    public double Carbohydrates { get; set; }
+    public double Fats { get; set; }
 
     public ICollection<MealFood> MealFoods { get; set; } = null!;
 }
